Store max_filesize as Float64 to suport Inf values.
Kind of sucks since file sizes are integers but you can't represent Inf 
as an integer. So either we use typemax(Int) as the default max_filesize 
or use Float64(Inf), which I think is cleaner.

+++ src/output_writers.jl
@@ -179,7 +179,7 @@ mutable struct JLD2OutputWriter{F, I, O, IF, IN} <: OutputWriter
        including :: IN
         previous :: Float64
             part :: Int
-    max_filesize :: Int
+    max_filesize :: Float64
            async :: Bool
            force :: Bool
          verbose :: Bool
